review fix

+++ backend/controllers/cards.js
@@ -1,6 +1,7 @@
 const cardModel = require('../models/card.js');
 const NotFoundError = require('../errors/not-found-err.js');
 const BadReqestError = require('../errors/bad-reqest.js');
+const ForbiddenError = require('../errors/forbidden-err.js');
 
 module.exports.getCards = (req, res, next) => {
   cardModel.find({})
@@ -20,8 +21,14 @@ module.exports.createCard = (req, res, next) => {
 };
 
 module.exports.deleteCard = (req, res, next) => {
-  cardModel.findByIdAndRemove(req.params.id)
-    .then(() => res.send({ message: 'deleted' }))
+  cardModel.findById(req.params.id)
+    .then((card) => {
+      if (card.owner.equals(req.user._id)) {
+        return cardModel.findByIdAndRemove(req.params.id)
+          .then(() => res.send({ message: 'deleted' }));
+      }
+      return next(new ForbiddenError('Forbidden'));
+    })
     .catch((err) => {
       if (err.name === 'CastError') {
         return next(new NotFoundError('Нет карточки с таким id'));
@@ -34,25 +41,24 @@ module.exports.putLike = (req, res, next) => {
   cardModel.findByIdAndUpdate(req.params.id,
     { $addToSet: { likes: req.user._id } },
     { new: true })
-    .then((card) => res.send(card))
-    .catch((err) => {
-      if (err.name === 'CastError') {
+    .then((card) => {
+      if (!card) {
         return next(new NotFoundError('Нет карточки с таким id'));
       }
-      return next();
-    });
+      return res.send(card);
+    })
+    .catch(() => next());
 };
 
 module.exports.deleteLike = (req, res, next) => {
   cardModel.findByIdAndUpdate(req.params.id,
     { $pull: { likes: req.user._id } },
     { new: true })
-    .orFail(() => new Error('NotFound'))
-    .then((card) => res.send(card))
-    .catch((err) => {
-      if (err.name === 'CastError') {
+    .then((card) => {
+      if (!card) {
         return next(new NotFoundError('Нет карточки с таким id'));
       }
-      return next();
-    });
+      return res.send(card);
+    })
+    .catch(() => next());
 };
